requirements.txt
pandas==2.2.3
streamlit==1.38.0
plotly-express==0.4.1
requests==2.32.3